Bugfix, creating a new Task when refreshing the page
